Add handle_books, handle_people, handle_rentals
